requirements.txt
tavily-python
google-generativeai
streamlit
python-dotenv